adding loading skeletons
