+++ ui/app/[locale]/(root_layout)/oratio/[id]/loading.tsx
@@ -1,0 +1,8 @@
+"use client"
+
+import React from "react";
+import SkeletonContent from "@/components/SkeletonContent";
+
+export default function Loading() {
+  return <SkeletonContent />
+}
