+++ ui/app/[locale]/(root_layout)/votive/loading.tsx
@@ -1,0 +1,8 @@
+"use client"
+
+import React from "react";
+import SkeletonIndex from "@/components/SkeletonIndex";
+
+export default function Loading() {
+  return <SkeletonIndex />
+}
